feat: 添加tabs删除事件支持，优化删除逻辑以避免无效操作，移除apidocBaseInfo中的spendTime字段

+++ packages/web/src/renderer/store/apidoc/tabs.ts
@@ -148,10 +148,12 @@ export const useApidocTas = defineStore('apidocTabs', () => {
     if (force) {
       ids.forEach((id) => {
         const deleteIndex = tabs.value[projectId].findIndex((tab) => tab._id === id);
-        deleteTabByIndex({
-          projectId,
-          deleteIndex,
-        })
+        if (deleteIndex !== -1) {
+          deleteTabByIndex({
+            projectId,
+            deleteIndex,
+          })
+        }
       })
       freshNewSeletedTab();
       return
